requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.29
pydantic==2.7.1
psycopg2-binary==2.9.2
python-dotenv==1.0.1


